pyproject.toml
[project]
name = "voice-assistance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "dspy>=3.0.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-genai>=1.39.1",
    "gtts>=2.5.4",
    "mem0ai>=0.1.116",
    "neo4j>=6.0.1",
    "openai>=2.0.0",
    "speechrecognition>=3.14.3",
]
